requirements.txt
psycopg2-binary==2.9.7
faker==19.3.0
pytest==7.4.2
python-dotenv==1.0.0
